# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OmPay API",
        "description": "API pour l'application OmPay - Système de paiement mobile sénégalais avec authentification double niveau",
        "contact": {
            "email": "contact@ompay.sn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur de développement"
        },
        {
            "url": "https://ompay-sa.onrender.com/api/v1",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/client/transactions/payment": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Paiement marchand",
                "description": "Effectue un paiement chez un marchand",
                "operationId": "payment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "code_marchand": {
                                        "description": "Code du marchand",
                                        "type": "string",
                                        "example": "M123456"
                                    },
                                    "numero_marchand": {
                                        "description": "Numéro de téléphone du marchand",
                                        "type": "string",
                                        "example": "+221781157773"
                                    },
                                    "montant": {
                                        "description": "Montant du paiement",
                                        "type": "integer",
                                        "example": 5000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paiement effectué",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paiement effectué avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction_id": {
                                                    "type": "string",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "paiement_marchand"
                                                },
                                                "identifiant_marchand": {
                                                    "type": "string",
                                                    "example": "M123456"
                                                },
                                                "montant": {
                                                    "type": "string",
                                                    "example": "-5000"
                                                },
                                                "solde_apres": {
                                                    "type": "integer",
                                                    "example": 95000
                                                },
                                                "date_creation": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-11-09T10:05:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Solde insuffisant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors du paiement"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/client/transactions/transfert": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Transfert entre clients OMPay",
                "description": "Effectue un transfert d'argent vers un autre client OMPay",
                "operationId": "transfert",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "numero_ompay": {
                                        "description": "Numéro OMPay du destinataire",
                                        "type": "string",
                                        "example": "+221781157773"
                                    },
                                    "montant": {
                                        "description": "Montant du transfert",
                                        "type": "integer",
                                        "example": 3000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfert effectué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transfert effectué avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "transaction_id": {
                                                    "type": "string",
                                                    "example": "550e8400-e29b-41d4-a716-446655440001"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "transfert"
                                                },
                                                "montant": {
                                                    "type": "integer",
                                                    "example": 3000
                                                },
                                                "frais": {
                                                    "type": "integer",
                                                    "example": 24
                                                },
                                                "emetteur": {
                                                    "properties": {
                                                        "numero_ompay": {
                                                            "type": "string",
                                                            "example": "+221781157772"
                                                        },
                                                        "solde_apres": {
                                                            "type": "integer",
                                                            "example": 92024
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "destinataire": {
                                                    "properties": {
                                                        "numero_ompay": {
                                                            "type": "string",
                                                            "example": "+221781157773"
                                                        },
                                                        "nom_prenom": {
                                                            "type": "string",
                                                            "example": "Moussa Diallo"
                                                        },
                                                        "montant_recu": {
                                                            "type": "integer",
                                                            "example": 3000
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "date_creation": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-11-09T10:05:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Solde insuffisant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur lors du transfert"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "title": "Réponse d'erreur",
                "description": "Réponse en cas d'erreur",
                "properties": {
                    "success": {
                        "description": "Statut de succès",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Message d'erreur",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT obtenu via l'authentification admin ou client",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}